contacts: CAF conflated "Delete" by adding multi-delete
When CAF added a multi-contact delete feature, rather than creating
a new string, they reused an existing menu id and string.

If you editing a contact, "menu_delete" deletes that contact.  If you
are in the list of contacts, "menu_delete" brings up a CAF people
for multi-contact deletion.

This is mildly confusing for users because having a menu item that
completely changes its functionality based on context is wonky.

If you ignore that, all well and good except for the fact that on
tablet layouts, both options are active at the same time.
Prior to this change, Delete would always invoke the multi-contact
delete and it was no longer possible to delete the contact that you
were viewing!

Break the conflation of these two completely different behaviours.

Change-Id: I096d275a7f8bd919908c202cfce3acbe75126840